feat: user login info added in sidebar
